Repoint global-jjb to GitHub mirror
The Linux Foundation has started mirroring global-jjb to GitHub. As
such, we need to update the submodule checkout to use this resource
instead of the direct upstream Gerrit repository as it will allow LF to
perform systems maintenance on Gerrit without disrupting the downstream
projects that depend on the submodule.

Issue: IN-10
Change-Id: I8246c6bca83bac192c6be32f93efd5abe3168509
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>