[FAB-6191] Support multiple versions of Go
v1.0.x (currently release branch) requires
Go 1.7.5 and v1.1 (currently master branch)
requires Go 1.9.  For the CI jobs which
exclusively use Docker, this is not an
issue.  But there are multiple jobs (e.g.
E2E and the jobs which run behave tests)
which use the Go binary installed on the
slave node itself.

FAB-6191 has the details, but basically
this change creates a new builder which
sets the Go env from a properties file
stored in the fabric repo.  This change
depends on FAB-6365.

Change-Id: I5a51a10703684997e9b574629c9b90b74baeaf00
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>
Signed-off-by: RameshBabu <rameshbabu.thoomu@gmail.com>