[FAB-6644] Support multiple Go ver for CA jobs
The release and master branches of fabric-ca currently
support different versions of Go.  In order to support
this, we need to be able to inject branch specific
properties into the fabric-ca jobs.  This is
accomplished by using a ci.properties file in each
branch.  For Go, we use the GO_VER property.

Change-Id: Ica7399f3b9a623ddfb2f418c2bf2557dce35db55
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>