FAB-7621 Conditional based trigger
trigger downstream jobs based on the conditional-step match.
Upstream job has to trigger downstream jobs based on the matched
conditional step. Ignore build if the change is WIP next Trigger
documentation build if the change is related to .rst and trigger
byfn checks if the change is other than above two.

Change-Id: I27dcea42e70a168261dcd3dcd8301c15599b95ef
Signed-off-by: rameshthoomu <rameshbabu.thoomu@gmail.com>