Use global-jjb rtd-verify job for docs builds
This patch updates the fabric project to use the global-jjb template
for handling Sphinx documentation builds. Details on the job
template can be found in the global-jjb documentation here [0] as
well as source code available here [1].

This job by default builds against new patches submitted or Gerrit
comment "recheck". Additionally only triggers if a file with the
pattern **/*.rst or **/conf.py is modified. These configurations
can be overrided via the following parameters:

- gerrit_verify_triggers
- gerrit_trigger_file_paths

Please refer to global-jjb documentation for more details on
specific parameter overrides that are possible with this job.

[0] http://global-jjb.releng.linuxfoundation.org/en/latest/jjb/lf-rtd-jobs.html
[1] https://github.com/lfit/releng-global-jjb/blob/master/jjb/lf-rtd-jobs.yaml

Depends-On: https://gerrit.hyperledger.org/r/17275
Change-Id: I5c6b83a815a4229210565bc3849d332c265e07b6
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>