Improve npm-install macro into a reusable macro
This improves the 'npm-install' macro in the fabric project to
become a more generic macro that allows projects to choose which
version NodeJS and NVM they wish to install. It introduces 2 new
macro variables:

- node-version
- nvm-version

Additionally this script updates the fabric-release-tests-{arch}
template to work with the new macro. By default the template will
install node-version 8.9.1 and nvm 0.33.2 as it did before but
can be overrided at the project declaration level by passing the
2 parameters mentioned above.

As a best practice we've moved the shell into it's own script away
from inlining at the JJB level. This allows us to take advantage
of editors that can read bash code as well as let's the ShellCheck
linter scan the code for code issues.

This patch conforms to the best practice guidelines as described:
http://global-jjb.releng.linuxfoundation.org/en/latest/best-practices.html

Change-Id: Ife114e809bbbe3f2c9fca19146e1436d6b345f81
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>