push thirdparty docker images for 1.0.x releases
We are using the same release scheme for 1.0 and 1.1, but the
scheme only works properly for 1.1 because of the fact that we
publish the kafka, couch, etc images from baseimage now.

As a work around, this patch pushes the docker images based on the
GERRIT_REFSPEC value. If the value is 1.0.x then it publishes thirdparty
images along with fabric images otherwise, will publish only fabric
docker images assuming thirdparty images are published from baseimage
repo.

Change-Id: I4c663cbb6692015affb8f9d155380446e8ae82c7
Signed-off-by: rameshthoomu <rameshbabu.thoomu@gmail.com>