FAB-7454 Clone the latest commit in merge jobs
The existing merge jobs always fetch the code from the gerrit refspec
instead of fetching the latest commit from the repository. This is same 
as what verify job is doing. To avoid duplicate verification of the 
patch set, this patch basically provides a way to clone the latest
commit from the repository.

Every merge job in production jenkins, will get the repository with 
"refs/heads/$GERRIT_BRANCH"

Change-Id: I559e36de8e1c5f1db2d435dfcd7125e103c22a6e
Signed-off-by: rameshthoomu <rameshbabu.thoomu@gmail.com>