Modify script in fabric-ca verify/merge jobs.
This change is to modify the builder script for
fabric-ca verify & merge jobs. Currently the
builder script runs hyperledger/fabric-ca-tools,
this has been changed to hyperledger/fabric-ca-fvt
which is the correct path.

Increase build timeout by 20 minutes for jobs in
fabric-sdk-java.

Increase build timeout to 60 minutes for jobs in
fabric-ca, wherever the tmeout is less than 60 mins.

Change-Id: Iaa8f6ff0c036fb708c6a994af5d828ee72270cc8
Signed-off-by: Sambhav Nidamarty <sambhavdutt@gmail.com>