[FAB-9277] Remove javaenv from CI scripts
Java chaincode is in the process of being
rewritten in order to be sully supported in a
future release.  The code has been move to its
own repo.  It was an experimental feature in
v1.0.x and v1.1.x and therefore there is no
reason to even publish images.  There are also
no actual tests which require it.

Change-Id: Ic09e575126dfa6d6afb8648a684a5f592e8925e9
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>