[FAB-6415] Remove JAR and whitespace checks
The JAR checks are a problem, as it stops us
checking in changes to valid JAR files such as
the Gradle wrapper JARs. The decision over
these files should be down to the maintainers,
not the CI pipeline.

The whitespace checks are also problematic,
and they should be merged into the repository
using a standard Java tool such as checkstyle.

Signed-off-by: Simon Stone <sstone1@uk.ibm.com>
Change-Id: I802e15ff280582f8408c88c9b60ef3cec8d00101